+++ libraries/oneview_resource_base.rb
@@ -20,7 +20,7 @@ module OneviewCookbook
       context.property :save_resource_info, [TrueClass, FalseClass, Array], default: context.node['oneview']['save_resource_info']
       context.property :api_version, Fixnum, default: context.node['oneview']['api_version']
       context.property :api_variant, [String, Symbol], default: context.node['oneview']['api_variant']
-      context.property :api_header_version, Fixnum # Used in a header for API requests
+      context.property :api_header_version, Fixnum # Overrides X-API-Version headers in API requests
     end
   end